refactor: organize the nivo / needs viz stuff
