+++ src/components/vis/needs-bar-chart/data-helpers.test.ts
@@ -21,7 +21,7 @@ import {
   selectQuarter,
   selectRegion,
   selectSubregion,
-} from './needs-helpers'
+} from './data-helpers'
 
 import { getNeed } from '../../../types/need.test-helpers'
 import {
