+++ src/components/vis/needs-bar-chart/data-helpers.ts
